requirements.txt
openai>=1.0.0,<2
requests==2.32.3
python-dotenv==1.0.1
python-slugify==8.0.4
